Update 10-courses.Rmd
I replaced the section for years2-6 with a single sentence on line 41. I also like the idea of including other milestones (e.g. comprehensive exams, committee meetings, ect.) that may influence when students plan to take courses. Gavin said he would add this type of info in one of his comments so I did not leave a placeholder in the text.